[Telink] Add RPC for ligt-switch app
